+++ Amazon/Pay/API/Constants/DisputeResolution.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Amazon\Pay\API\Constants;
+
+class DisputeResolution {
+    public const BUYER_WON = "BuyerWon";
+    public const MERCHANT_WON = "MerchantWon";
+    public const NO_FAULT = "NoFault";
+}