+++ Amazon/Pay/API/Constants/DisputeState.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Amazon\Pay\API\Constants;
+
+class DisputeState {
+    public const UNDER_REVIEW = "UnderReview";
+    public const ACTION_REQUIRED = "ActionRequired";
+    public const RESOLVED = "Resolved";
+    public const CLOSED = "Closed";
+}