+++ Amazon/Pay/API/Constants/EvidenceType.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Amazon\Pay\API\Constants;
+
+class EvidenceType {
+    public const PRODUCT_DESCRIPTION = "ProductDescription";
+    public const RECEIPT = "Receipt";
+    public const CANCELLATION_POLICY = "CancellationPolicy";
+    public const CUSTOMER_SIGNATURE = "CustomerSignature";
+    public const TRACKING_NUMBER = "TrackingNumber";
+    public const OTHER = "Other";
+}